prepare publishing version 29.0.0-rc.1

+++ projects/ngx-extended-pdf-viewer/types/src/display/obj_bin_transform_display.d.ts
@@ -46,7 +46,7 @@ export class PatternInfo {
     buffer: any;
     view: DataView<any>;
     data: Uint8Array<any>;
-    getIR(): (string | number | number[] | Uint8Array<any> | Float32Array<any> | null)[] | (string | number | number[] | (string | number)[][] | null)[];
+    getIR(): (string | number | number[] | Float32Array<any> | Uint8Array<any> | null)[] | (string | number | number[] | (string | number)[][] | null)[];
 }
 export class SystemFontInfo {
     constructor(buffer: any);
